El punto y coma

+++ controlador.php
@@ -585,7 +585,7 @@
 
             $id = $_REQUEST["id"];
 
-            if ($ajax != null; $this->seguridad->esAdmin() || $this->seguridad->get('id') == $id) {
+            if ($ajax != null && ($this->seguridad->esAdmin() || $this->seguridad->get('id') == $id)) {
                 echo $this->reserva->delete($id);
             } else {
                 $this->reserva->delete($id);
